prism/backend/requirements.txt
black==26.1.0
click==8.3.1
customtkinter==5.2.2
darkdetect==0.8.0
fastapi==0.111.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pytokens==0.4.0
uvicorn==0.30.0
google-generativeai
python-dotenv
